BandCamp Album Downloader
The "Init.bat" script launches launches Powershell and skips the
execution policy check, then launches the BC_Down powershell script.
This prompts user for the web address of the album on Bandcamp, asks
where to store the album (as a sub folder of the Public Desktop\BandCamp
Dowloaded folder), and then if the user wants to download another.
Currently, error checking is minimal.  Recently added the ability to
replace special chars that Windows won't allow in filenames with an
underscore.